Backend/requirements.txt
# /backend/requirements.txt
flask==2.3.3
authlib==1.3.1 #[?] For Oauth2
pyjwt==2.8.0 #[?] For JWT
flask-sqlalchemy==3.0.5 #[?] For SQL
psycopg2-binary==2.9.9 #[?]For PostgreSQL
flask_cors
requests
python-dotenv